Refactor container maintenance code
Move maintenance code to dedicated file and improve code layout and
format, splitting functions and include debugging message using default
logging system
